\name{allElem}
\alias{allElem}
\title{Get all permutation/combination for a permutation iterator}
\usage{
allElem(obj, ...)

\method{allElem}{piterator}(I)
}
\arguments{
  \item{I}{a permutation/combination iterator}
}
\value{
next permutation/combination sequence for the iterator
\code{I}
}
\description{
Get all permutation/combination for a permutation iterator
}

