\name{getAll}
\alias{getAll}
\title{Get all permutation/combination for a iterator}
\usage{
getAll(I)
}
\arguments{
  \item{I}{a permutation/combination iterator}
}
\value{
next permutation/combination sequence for the iterator
\code{I}
}
\description{
Get all permutation/combination for a iterator
}

