% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{getcurrent}
\alias{getcurrent}
\title{Get the current element of a iterator}
\usage{
getcurrent(I)
}
\arguments{
\item{I}{a permutation/combination iterator}
}
\value{
current element of a iterator
}
\description{
Get the current element of a iterator
}

