% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{getlength}
\alias{getlength}
\title{Get the length for a iterator}
\usage{
getlength(I)
}
\arguments{
\item{I}{a permutations/combinations iterator
an integer}
}
\description{
Get the length for a iterator
}

