\name{getNext}
\alias{getNext}
\title{Get the next permutation(s)/combination(s) for a iterator}
\usage{
getNext(I, d = 1)
}
\arguments{
  \item{I}{a permutation(s)/combination(s) iterator}

  \item{d}{number of permutation(s)/combination(s) wanted,
  default to 1}
}
\value{
next \code{d} permutation(s)/combination(s) sequence for
the iterator \code{I}
}
\description{
Get the next permutation(s)/combination(s) for a iterator
}

