% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{np_multiset}
\alias{np_multiset}
\title{Calculate the number of r-permutations of a multiset}
\usage{
np_multiset(f, r, bigz = FALSE)
}
\arguments{
\item{f}{the frequencies of the mutliset}

\item{r}{the number of object drawn from the multiset}

\item{bigz}{use gmp's Big Interger}
}
\value{
the number of r-permutations (Big Integer from gmp)
}
\description{
Calculate the number of r-permutations of a multiset
}
\examples{
x = c("a","a","b")
# possible permutations of size 2 are "aa", "ab" and "ba".
np_multiset(table(x), 2) # = 3
}

