\name{setgen}
\alias{setgen}
\title{Set generator for permuations or combinations}
\usage{
setgen(type, n = NULL, x = NULL, r = n, replace = FALSE,
  multiset = anyDuplicated(x))
}
\arguments{
  \item{type}{type}
}
\value{
a generator objector
}
\description{
Set generator for permuations or combinations
}

