\name{getall}
\alias{getall}
\title{Get all permutation/combination for a iterator}
\usage{
getall(I)
}
\arguments{
  \item{I}{a permutation/combination iterator}
}
\value{
next permutation/combination sequence for the iterator
\code{I}
}
\description{
Get all permutation/combination for a iterator
}

