% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iterator.R
\name{getlength}
\alias{getlength}
\title{Get the length for a iterator}
\usage{
getlength(I)
}
\arguments{
\item{I}{a permutation(s)/combination(s) iterator
an integer}
}
\description{
Get the length for a iterator
}

