% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{getnext}
\alias{getnext}
\title{Get the next permutation(s)/combination(s) for a iterator}
\usage{
getnext(I, d = 1, drop = TRUE)
}
\arguments{
\item{I}{a permutation/combination iterator}

\item{d}{number of permutation(s)/combination(s) wanted, default to 1}

\item{drop}{if \code{d} is 1, drop simplify to vector if possible, default to \code{TRUE}.}
}
\value{
next \code{d} permutation(s)/combination(s) sequence for the iterator \code{I}
}
\description{
Get the next permutation(s)/combination(s) for a iterator
}

