% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multichoose}
\alias{multichoose}
\title{Calculate mulinomial coefficient}
\usage{
multichoose(n)
}
\arguments{
\item{n}{a vector of group sizes}
}
\value{
multinomial coefficient
}
\description{
This function calculate the multinomial coefficient
\deqn{\frac{(\sum n_j)!}{\prod n_j!}.}{(\sum n_j)! / \prod n_j!.}
where \eqn{n_j}{n_j}'s are the number of multiplicities in the multiset.
}
\details{
The result is not reliable if the retunred value is too large (around 2^53) due to limition of integers.
}
\examples{
# (3+1+1)!/ (3! 1! 1!) = 20
multichoose(c(3,1,1))
}

