% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{multichoose}
\alias{multichoose}
\title{Calculate multinomial coefficient}
\usage{
multichoose(n, bigz = FALSE)
}
\arguments{
\item{n}{a vector of group sizes}

\item{bigz}{use gmp's Big Interger}
}
\value{
multinomial coefficient
}
\description{
This function calculates the multinomial coefficient
\deqn{\frac{(\sum n_j)!}{\prod n_j!}.}{(\sum n_j)! / \prod n_j!.}
where \eqn{n_j}{n_j}'s are the number of multiplicities in the multiset.
}
\examples{
# (3+1+1)!/ (3! 1! 1!) = 20
multichoose(c(3,1,1))
}

