% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nc_multiset}
\alias{nc_multiset}
\title{Calculate the number of r-combinations of a multiset}
\usage{
nc_multiset(f, r, bigz = FALSE)
}
\arguments{
\item{f}{the frequencies of the mutliset}

\item{r}{the number of object drawn from the multiset}

\item{bigz}{use gmp's Big Interger}
}
\value{
the number of combinations (Big Integer from gmp)
}
\description{
Calculate the number of r-combinations of a multiset
}
\examples{
x <- c("a","a","b")
# possible combinations of size 2 are "aa" and "ab".
nc_multiset(table(x), 2) # <- 2
}

