\name{nextElem}
\alias{nextElem}
\title{Get the next permutation/combination for a iterator}
\usage{
nextElem(I, ...)

\method{nextElem}{piterator}(I)

\method{nextElem}{citerator}(I)
}
\arguments{
  \item{I}{a permutation/combination iterator}
}
\value{
next permutation/combination sequence for the iterator
\code{I}
}
\description{
Get the next permutation/combination for a iterator
}

